Fixed potential conversion errors and other compiler warnings.

+++ include/robot_model/HelperFunctions.hpp
@@ -17,7 +17,7 @@ namespace robot_model
     KDL::Rotation toKdl(urdf::Rotation r);
     KDL::Frame toKdl(urdf::Pose p);
     urdf::Pose toURDFPose(KDL::Frame frame);
-    void KDLFrameToEigenMatrix(KDL::Frame &frame,Eigen::Isometry3f &transform);
+    void KDLFrameToEigenMatrix(KDL::Frame &frame,Eigen::Isometry3d &transform);
 }
 
 #endif
